refactored EventList and App components to use composables -Refactored the EventList and App components to utilize composables for managing state and logic. -Moved booking and event registration logic into reusable composables (useBookings) for better code organization and reuse. -Improved maintainability by separating concerns and encapsulating logic in composables.
